fix: consolidate session save sources for pkg/service file-count limit
